Fix two corner cases which result in deadlocks.
This patch attempts to fix two corner cases which can result in hhpc
becoming deadlocked, <https://github.com/aktau/hhpc/issues/9>.

The first checks for the case where nanosleep() is interrupted which
would set working to 0 and as hhpc never recovers from this case, the
program will lockup indefinitely.  Instead we resume where we left of if
encountering EINTR.

The second fix is a bit of a hack; on rare occasions the select() call
in waitForMotion can seemingly block and practically never recover (I
didn't wait for more than one hour).

Instead I added a timeout to the select call, based on the gIdleTimeout
value if it is greater than 1, otherwise it uses a 3 second delay.
(This implies you can set a timeout of 2 seconds and have a 2 second
timeout.)

This doesn't really solve the issue of why select is getting stuck but
does allow hhpc to recover and prevent it from holding the mouse
hostage.